+++ src/Phan/Issue.php
@@ -68,6 +68,7 @@ class Issue
     const UndeclaredTypeProperty    = 'PhanUndeclaredTypeProperty';
     const UndeclaredTypeThrowsType  = 'PhanUndeclaredTypeThrowsType';
     const UndeclaredVariable        = 'PhanUndeclaredVariable';
+    const UndeclaredThis            = 'PhanUndeclaredThis';
     const UndeclaredVariableDim     = 'PhanUndeclaredVariableDim';
     const UndeclaredVariableAssignOp = 'PhanUndeclaredVariableAssignOp';
     const UndeclaredClassInCallable = 'PhanUndeclaredClassInCallable';
@@ -907,6 +908,14 @@ class Issue
                 "Variable \${VARIABLE} is undeclared",
                 self::REMEDIATION_B,
                 11018
+            ),
+            new Issue(
+                self::UndeclaredThis,
+                self::CATEGORY_UNDEFINED,
+                self::SEVERITY_CRITICAL,
+                "Variable \${VARIABLE} is undeclared",
+                self::REMEDIATION_B,
+                11046
             ),
             new Issue(
                 self::UndeclaredTypeParameter,